server/src/requirements.txt
bcrypt
jwt
pydantic
sqlalchemy
fastapi[all]
python-jose